+++ rooms/urls.py
@@ -1,9 +1,9 @@
 from django.urls import path
 
-from rooms.views.view_v0_room import RoomPhotos
 from rooms.views.view_v1_amenity import AmenitiesListAPI, RoomAmenitiesAPI
 from rooms.views.view_v1_room import (
     RoomCreateAPI,
+    RoomCreatePhotosAPI,
     RoomDeleteAPI,
     RoomDetailAPI,
     RoomsListAPI,
@@ -17,6 +17,6 @@ urlpatterns = [
     path("v1/update/<int:room_id>", RoomUpdateAPI.as_view()),
     path("v1/delete/<int:room_id>", RoomDeleteAPI.as_view()),
     path("v1/<int:room_id>/amenities", RoomAmenitiesAPI.as_view()),
-    path("v0/<int:pk>/photos", RoomPhotos.as_view()),
+    path("v1/<int:room_id>/photos", RoomCreatePhotosAPI.as_view()),
     path("v1/amenities/list", AmenitiesListAPI.as_view()),
 ]
