--- rooms/views/view_v0_room.py
@@ -1,100 +1,0 @@
-from django.conf import settings
-from drf_yasg import openapi
-from drf_yasg.utils import swagger_auto_schema
-from rest_framework import status
-
-from rest_framework.permissions import IsAuthenticatedOrReadOnly
-from rest_framework.views import APIView
-from rest_framework.status import HTTP_204_NO_CONTENT
-from rest_framework.response import Response
-from rest_framework.exceptions import (
-    NotFound,
-    ParseError,
-    PermissionDenied,
-)
-from rooms.models import Amenity, Room
-from categories.models import Category
-from rooms import serializers
-from reviews.serializers import ReviewSerializer
-from medias.serializers import PhotoSerializer
-from rooms.selectors.selector_v1_room import RoomSelector
-
-
-class RoomReviews(APIView):
-
-    permission_classes = [IsAuthenticatedOrReadOnly]
-
-    def get_object(self, pk):
-        try:
-            return Room.objects.get(pk=pk)
-        except Room.DoesNotExist:
-            raise NotFound
-
-    def get(self, request, pk):
-        try:
-            page = request.query_params.get("page", 1)
-            page = int(page)
-        except ValueError:
-            page = 1
-        page_size = settings.PAGE_SIZE
-        start = (page - 1) * page_size
-        end = start + page_size
-        room = self.get_object(pk)
-        serializer = ReviewSerializer(
-            room.reviews.all()[start:end],
-            many=True,
-        )
-        return Response(serializer.data)
-
-    def post(self, request, pk):
-        serializer = ReviewSerializer(data=request.data)
-        if serializer.is_valid():
-            review = serializer.save(
-                user=request.user,
-                room=self.get_object(pk),
-            )
-            serializer = ReviewSerializer(review)
-            return Response(serializer.data)
-
-
-class RoomAmenities(APIView):
-    def get(self, request, pk):
-        selector = RoomSelector(request)
-        room = selector.get_room(pk)
-        try:
-            page = request.query_params.get("page", 1)
-            page = int(page)
-        except ValueError:
-            page = 1
-        page_size = settings.PAGE_SIZE
-        start = (page - 1) * page_size
-        end = start + page_size
-
-        serializer = serializers.AmenitySerializer(
-            room.amenities.all()[start:end],
-            many=True,
-        )
-        return Response(serializer.data)
-
-
-class RoomPhotos(APIView):
-
-    permission_classes = [IsAuthenticatedOrReadOnly]
-
-    def get_object(self, pk):
-        try:
-            return Room.objects.get(pk=pk)
-        except Room.DoesNotExist:
-            raise NotFound
-
-    def post(self, request, pk):
-        room = self.get_object(pk)
-        if request.user != room.owner:
-            raise PermissionDenied
-        serializer = PhotoSerializer(data=request.data)
-        if serializer.is_valid():
-            photo = serializer.save(room=room)
-            serializer = PhotoSerializer(photo)
-            return Response(serializer.data)
-        else:
-            return Response(serializer.errors)
