+++ rooms/views/view_v1_amenity.py
@@ -1,12 +1,11 @@
 from drf_yasg import openapi
 from drf_yasg.utils import swagger_auto_schema
-
 from rest_framework import status
 from rest_framework.request import Request
-from rest_framework.views import APIView
 from rest_framework.response import Response
+from rest_framework.views import APIView
 
-from rooms.models import Amenity
+from rooms.models.room import Amenity
 from rooms.selectors.selector_v1_room import RoomSelector
 from rooms.serializers import AmenityListSerializer
 from utils.paginations import RoomAmenitiesPagination
