+++ rooms/views/view_v1_room.py
@@ -1,24 +1,26 @@
 from drf_yasg import openapi
 from drf_yasg.utils import swagger_auto_schema
 from rest_framework import status
-
-from rest_framework.permissions import IsAuthenticatedOrReadOnly, IsAuthenticated
+from rest_framework.permissions import IsAuthenticated, IsAuthenticatedOrReadOnly
 from rest_framework.request import Request
-from rest_framework.views import APIView
-from rest_framework.status import HTTP_204_NO_CONTENT
 from rest_framework.response import Response
+from rest_framework.status import HTTP_204_NO_CONTENT
+from rest_framework.views import APIView
+
+from medias.serializers import PhotoCreateSerializer
+from medias.services.service_v1_photo import PhotoService
 from rooms import serializers
 from rooms.selectors.selector_v1_room import RoomSelector
 from rooms.serializers import (
-    RoomListOutputSerializer,
     RoomDetailOutputSerializer,
+    RoomListOutputSerializer,
     RoomUpdateInputSerializer,
 )
-from rooms.services.service_v1_room import RoomService, RoomDeleteService
+from rooms.services.service_v1_room import RoomDeleteService, RoomService
+from utils.exceptions.exception import RoomExceptions
 
 
 class RoomsListAPI(APIView):
-
     permission_classes = [IsAuthenticatedOrReadOnly]
 
     @swagger_auto_schema(
@@ -31,7 +33,6 @@ class RoomsListAPI(APIView):
         },
     )
     def get(self, request: Request):
-
         rooms = RoomSelector(request).get_all_rooms()
         serializer = RoomListOutputSerializer(rooms, many=True)
 
@@ -39,7 +40,6 @@ class RoomsListAPI(APIView):
 
 
 class RoomCreateAPI(APIView):
-
     permission_classes = [IsAuthenticated]
 
     @swagger_auto_schema(
@@ -60,7 +60,6 @@ class RoomCreateAPI(APIView):
 
 
 class RoomDetailAPI(APIView):
-
     permission_classes = [IsAuthenticatedOrReadOnly]
 
     @swagger_auto_schema(
@@ -78,7 +77,6 @@ class RoomDetailAPI(APIView):
 
 
 class RoomUpdateAPI(APIView):
-
     # TODO: Owner Permission
     permission_classes = [IsAuthenticated]
 
@@ -103,7 +101,6 @@ class RoomUpdateAPI(APIView):
 
 
 class RoomDeleteAPI(APIView):
-
     # TODO: Owner Permission
     permission_classes = [IsAuthenticated]
 
@@ -122,3 +119,28 @@ class RoomDeleteAPI(APIView):
         service.delete_room()
 
         return Response(status=HTTP_204_NO_CONTENT)
+
+
+class RoomCreatePhotosAPI(APIView):
+    permission_classes = [IsAuthenticatedOrReadOnly]
+
+    @swagger_auto_schema(
+        operation_summary="V1 Room Photo Create API",
+        operation_description="Photo 정보를 받아 저장후 해당 Room에 매핑",
+        responses={
+            status.HTTP_200_OK: openapi.Response("매핑 완료"),
+        },
+    )
+    def post(self, request: Request, room_id: int) -> Response:
+        room = RoomSelector(request).get_room(room_id)
+
+        if request.user != room.owner:
+            raise RoomExceptions.UserIsNotOwner
+
+        input_serializer = PhotoCreateSerializer(data=request.data)
+        input_serializer.is_valid(raise_exception=True)
+
+        service = PhotoService(request, input_serializer.validated_data)
+        photo = service.create_photo(room)
+        output_serializer = PhotoCreateSerializer(photo)
+        return Response(output_serializer.data, status=status.HTTP_200_OK)
